+++ packages/common/src/downloads/index.ts
@@ -1,1 +1,6 @@
 export * from "./build-existing-exports-portal-query";
+export * from "./types";
+export * from "./fetchDownloadFileUrl";
+export * from "./fetchDownloadFormats";
+export * from "./canUseCreateReplica";
+export * from "./canUseHubDownloadApi";